# docs/contracts/openapi-v1.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "SOUL V1 API",
    "version": "1.0.0",
    "description": "Machine-readable companion to docs/FLUTTER_API_HANDOFF.md."
  },
  "servers": [
    {
      "url": "http://localhost/api/v1",
      "description": "Local development; replace the origin per environment"
    }
  ],
  "paths": {
...
    "/catalogs/profile": {
      "get": {
        "operationId": "api.v1.catalogs.profile",
        "summary": "Localized interests, traits and help categories",
        "tags": [
          "Catalogs"
        ],
        "responses": {
          "401": {
            "$ref": "#/components/responses/Error"
          },
          "422": {
            "$ref": "#/components/responses/Error"
          },
          "429": {
            "$ref": "#/components/responses/Error"
          },
          "2XX": {
            "$ref": "#/components/responses/Success"
          }
        }
      }
    },
...
    "/auth/devices/{session}": {
      "delete": {
        "operationId": "api.v1.auth.devices.destroy",
        "summary": "Remotely sign out one owned device session",
        "tags": [
          "Auth"
        ],
        "parameters": [
          {
            "name": "session",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "401": {
            "$ref": "#/components/responses/Error"
          },
          "422": {
            "$ref": "#/components/responses/Error"
          },
          "429": {
            "$ref": "#/components/responses/Error"
          },
          "2XX": {
            "$ref": "#/components/responses/Success"
          }
        }
      }
    },
...
    "/profiles/{profile}/like": {
      "put": {
        "operationId": "api.v1.likes.update",
        "summary": "Accept or decline a pending Like",
        "tags": [
          "Likes"
        ],
        "parameters": [
          {
            "name": "profile",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object"
              },
              "example": {
                "decision": "accept"
              }
            }
          }
        },
        "responses": {
          "401": {
            "$ref": "#/components/responses/Error"
          },
          "422": {
            "$ref": "#/components/responses/Error"
          },
          "429": {
            "$ref": "#/components/responses/Error"
          },
          "2XX": {
            "$ref": "#/components/responses/Success"
          }
        }
      },
      "delete": {
        "operationId": "api.v1.likes.destroy",
        "summary": "Withdraw a pending outgoing Like",
        "tags": [
          "Likes"
        ],
        "parameters": [
          {
            "name": "profile",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "401": {
            "$ref": "#/components/responses/Error"
          },
          "422": {
            "$ref": "#/components/responses/Error"
          },
          "429": {
            "$ref": "#/components/responses/Error"
          },
          "2XX": {
            "$ref": "#/components/responses/Success"
          }
        }
      }
    },
...
    "/matches/{match}": {
      "delete": {
        "operationId": "api.v1.matches.destroy",
        "summary": "Idempotent unmatch",
        "tags": [
          "Interactions"
        ],
        "parameters": [
          {
            "name": "match",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "401": {
            "$ref": "#/components/responses/Error"
          },
          "422": {
            "$ref": "#/components/responses/Error"
          },
          "429": {
            "$ref": "#/components/responses/Error"
          },
          "2XX": {
            "$ref": "#/components/responses/Success"
          }
        }
      }
    },
...
    "/private-photo-access/{accessRequest}": {
      "put": {
        "operationId": "api.v1.private-photo-access.update",
        "summary": "Owner approves or rejects",
        "tags": [
          "Private-photo-access"
        ],
        "parameters": [
          {
            "name": "accessRequest",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object"
              },
              "example": {
                "decision": "approve"
              }
            }
          }
        },
        "responses": {
          "401": {
            "$ref": "#/components/responses/Error"
          },
          "422": {
            "$ref": "#/components/responses/Error"
          },
          "429": {
            "$ref": "#/components/responses/Error"
          },
          "2XX": {
            "$ref": "#/components/responses/Success"
          }
        }
      },
      "delete": {
        "operationId": "api.v1.private-photo-access.destroy",
        "summary": "Owner revokes access",
        "tags": [
          "Private-photo-access"
        ],
        "parameters": [
          {
            "name": "accessRequest",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "401": {
            "$ref": "#/components/responses/Error"
          },
          "422": {
            "$ref": "#/components/responses/Error"
          },
          "429": {
            "$ref": "#/components/responses/Error"
          },
          "2XX": {
            "$ref": "#/components/responses/Success"
          }
        }
      }
    },
...
    "/devices/{device}": {
      "delete": {
        "operationId": "api.v1.devices.destroy",
        "summary": "Revoke owned device",
        "tags": [
          "Devices"
        ],
        "parameters": [
          {
            "name": "device",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "401": {
            "$ref": "#/components/responses/Error"
          },
          "422": {
            "$ref": "#/components/responses/Error"
          },
          "429": {
            "$ref": "#/components/responses/Error"
          },
          "2XX": {
            "$ref": "#/components/responses/Success"
          }
        }
      }
    },
...
    "/events/{event}/registration": {
      "post": {
        "operationId": "api.v1.events.registration.store",
        "summary": "Capacity-safe idempotent join",
        "tags": [
          "Events"
        ],
        "parameters": [
          {
            "name": "event",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "401": {
            "$ref": "#/components/responses/Error"
          },
          "422": {
            "$ref": "#/components/responses/Error"
          },
          "429": {
            "$ref": "#/components/responses/Error"
          },
          "2XX": {
            "$ref": "#/components/responses/Success"
          }
        }
      },
      "delete": {
        "operationId": "api.v1.events.registration.destroy",
        "summary": "Idempotent leave",
        "tags": [
          "Events"
        ],
        "parameters": [
          {
            "name": "event",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "401": {
            "$ref": "#/components/responses/Error"
          },
          "422": {
            "$ref": "#/components/responses/Error"
          },
          "429": {
            "$ref": "#/components/responses/Error"
          },
          "2XX": {
            "$ref": "#/components/responses/Success"
          }
        }
      }
    },
...
    "/privacy/deletion": {
      "post": {
        "operationId": "api.v1.privacy.deletion.store",
        "summary": "Schedule deletion with confirmation",
        "tags": [
          "Privacy"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object"
              },
              "example": {
                "confirmation": "DELETE MY ACCOUNT"
              }
            }
          }
        },
        "responses": {
          "401": {
            "$ref": "#/components/responses/Error"
          },
          "422": {
            "$ref": "#/components/responses/Error"
          },
          "429": {
            "$ref": "#/components/responses/Error"
          },
          "2XX": {
            "$ref": "#/components/responses/Success"
          }
        }
      },
      "get": {
        "operationId": "api.v1.privacy.deletion.show",
        "summary": "Resume scheduled-deletion state",
        "tags": [
          "Privacy"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "401": {
            "$ref": "#/components/responses/Error"
          },
          "422": {
            "$ref": "#/components/responses/Error"
          },
          "429": {
            "$ref": "#/components/responses/Error"
          },
          "2XX": {
            "$ref": "#/components/responses/Success"
          }
        }
      },
      "delete": {
        "operationId": "api.v1.privacy.deletion.destroy",
        "summary": "Cancel inside grace period",
        "tags": [
          "Privacy"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "401": {
            "$ref": "#/components/responses/Error"
          },
          "422": {
            "$ref": "#/components/responses/Error"
          },
          "429": {
            "$ref": "#/components/responses/Error"
          },
          "2XX": {
            "$ref": "#/components/responses/Success"
          }
        }
      }
    },
...
    "/admin/admins/{admin}": {
      "delete": {
        "operationId": "api.v1.admin.admins.destroy",
        "summary": "Remove another admin's access safely",
        "tags": [
          "Admin"
        ],
        "parameters": [
          {
            "name": "admin",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "security": [
          {
            "adminSession": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object"
              },
              "example": {
                "reason": "Admin left operations"
              }
            }
          }
        },
        "responses": {
          "401": {
            "$ref": "#/components/responses/Error"
          },
          "422": {
            "$ref": "#/components/responses/Error"
          },
          "429": {
            "$ref": "#/components/responses/Error"
          },
          "2XX": {
            "$ref": "#/components/responses/Success"
          }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "Sanctum token"
      },
      "adminSession": {
        "type": "apiKey",
        "in": "cookie",
        "name": "laravel_session"
      },
      "cloudinarySignature": {
        "type": "apiKey",
        "in": "header",
        "name": "X-Cld-Signature"
      }
    },
    "schemas": {
      "SuccessEnvelope": {
        "type": "object",
        "required": [
          "success",
          "message",
          "data",
          "meta"
        ],
        "properties": {
          "success": {
            "const": true
          },
          "message": {
            "type": "string"
          },
          "data": {},
          "meta": {
            "type": "object",
            "properties": {
              "request_id": {
                "type": "string"
              }
            }
          }
        }
      },
      "ErrorEnvelope": {
        "type": "object",
        "required": [
          "success",
          "error",
          "meta"
        ],
        "properties": {
          "success": {
            "const": false
          },
          "error": {
            "type": "object",
            "required": [
              "code",
              "message",
              "details"
            ],
            "properties": {
              "code": {
                "type": "string"
              },
              "message": {
                "type": "string"
              },
              "details": {
                "type": "object"
              }
            }
          },
          "meta": {
            "type": "object",
            "properties": {
              "request_id": {
                "type": "string"
              }
            }
          }
        }
      }
    },
    "responses": {
      "Success": {
        "description": "Successful SOUL response",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/SuccessEnvelope"
            }
          }
        }
      },
      "Error": {
        "description": "Standard SOUL error response",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ErrorEnvelope"
            }
          }
        }
      }
    }
  }
}